+++ src/psycop_model_training/model_eval/base_artifacts/base_artifact_generator.py
@@ -5,22 +5,20 @@ from psycop_model_training.model_eval.base_artifacts.plots.feature_importance im
     plot_feature_importances,
 )
 from psycop_model_training.model_eval.base_artifacts.plots.performance_by_age import (
-    plot_performance_by_age,
+    plot_roc_auc_by_age,
 )
 from psycop_model_training.model_eval.base_artifacts.plots.performance_over_time import (
-    plot_auc_by_time_from_first_visit,
     plot_metric_by_calendar_time,
-    plot_metric_by_cyclic_time,
-    plot_metric_by_time_until_diagnosis,
     plot_recall_by_calendar_time,
+    plot_roc_auc_by_cyclic_time,
+    plot_roc_auc_by_time_from_first_visit,
+    plot_sensitivity_by_time_until_diagnosis,
+    plot_sensitivity_by_time_to_outcome_heatmap,
 )
 from psycop_model_training.model_eval.base_artifacts.plots.precision_recall import (
     plot_precision_recall,
 )
 from psycop_model_training.model_eval.base_artifacts.plots.roc_auc import plot_auc_roc
-from psycop_model_training.model_eval.base_artifacts.plots.sens_over_time import (
-    plot_sensitivity_by_time_to_outcome_heatmap,
-)
 from psycop_model_training.model_eval.base_artifacts.plots.time_from_first_positive_to_event import (
     plot_time_from_first_positive_to_event,
 )
@@ -39,7 +37,6 @@ from psycop_model_training.model_eval.dataclasses import (
     EvalDataset,
     PipeMetadata,
 )
-from psycop_model_training.utils.utils import positive_rate_to_pred_probs
 from sklearn.metrics import recall_score
 
 
@@ -61,23 +58,18 @@ class BaseArtifactGenerator:
 
     def create_base_plot_artifacts(self) -> list[ArtifactContainer]:
         """A collection of plots that are always generated."""
-        pred_proba_percentiles = positive_rate_to_pred_probs(
-            pred_probs=self.eval_ds.y_hat_probs,
-            positive_rate_thresholds=self.cfg.eval.positive_rate_thresholds,
-        )
-
         lookahead_bins = self.cfg.eval.lookahead_bins
 
         return [
-            ArtifactContainer(
-                label="sensitivity_by_time_by_threshold",
-                artifact=plot_sensitivity_by_time_to_outcome_heatmap(
-                    eval_dataset=self.eval_ds,
-                    pred_proba_thresholds=pred_proba_percentiles,
-                    bins=lookahead_bins,
-                    save_path=self.save_dir / "sensitivity_by_time_by_threshold.png",
-                ),
-            ),
+            # ArtifactContainer(
+            #     label="sensitivity_by_time_by_threshold",
+            #     artifact=plot_sensitivity_by_time_to_outcome_heatmap(
+            #         eval_dataset=self.eval_ds,
+            #         pred_proba_thresholds=pred_proba_percentiles,
+            #         bins=lookahead_bins,
+            #         save_path=self.save_dir / "sensitivity_by_time_by_threshold.png",
+            #     ),
+            # ),
             # ArtifactContainer(
             #     label="auc_by_time_from_first_visit",
             #     artifact=plot_auc_by_time_from_first_visit(
@@ -95,7 +87,7 @@ class BaseArtifactGenerator:
             ),
             ArtifactContainer(
                 label="auc_by_hour_of_day",
-                artifact=plot_metric_by_cyclic_time(
+                artifact=plot_roc_auc_by_cyclic_time(
                     eval_dataset=self.eval_ds,
                     bin_period="H",
                     save_path=self.save_dir / "auc_by_hour_of_day.png",
@@ -103,7 +95,7 @@ class BaseArtifactGenerator:
             ),
             ArtifactContainer(
                 label="auc_by_day_of_week",
-                artifact=plot_metric_by_cyclic_time(
+                artifact=plot_roc_auc_by_cyclic_time(
                     eval_dataset=self.eval_ds,
                     bin_period="D",
                     save_path=self.save_dir / "auc_by_day_of_week.png",
@@ -111,7 +103,7 @@ class BaseArtifactGenerator:
             ),
             ArtifactContainer(
                 label="auc_by_month_of_year",
-                artifact=plot_metric_by_cyclic_time(
+                artifact=plot_roc_auc_by_cyclic_time(
                     eval_dataset=self.eval_ds,
                     bin_period="M",
                     save_path=self.save_dir / "auc_by_month_of_year.png",
@@ -126,10 +118,10 @@ class BaseArtifactGenerator:
             ),
             ArtifactContainer(
                 label="recall_by_time_to_diagnosis",
-                artifact=plot_metric_by_time_until_diagnosis(
+                artifact=plot_sensitivity_by_time_until_diagnosis(
                     eval_dataset=self.eval_ds,
                     bins=lookahead_bins,
-                    metric_fn=recall_score,
+                    roc_auc_score=recall_score,
                     y_title="Sensitivity (recall)",
                     save_path=self.save_dir / "recall_by_time_to_diagnosis.png",
                 ),
@@ -138,14 +130,12 @@ class BaseArtifactGenerator:
                 label="performance_by_threshold",
                 artifact=generate_performance_by_positive_rate_table(
                     eval_dataset=self.eval_ds,
-                    pred_proba_thresholds=pred_proba_percentiles,
-                    positive_rate_thresholds=self.cfg.eval.positive_rate_thresholds,
                     output_format="df",
                 ),
             ),
             ArtifactContainer(
                 label="performance_by_age",
-                artifact=plot_performance_by_age(
+                artifact=plot_roc_auc_by_age(
                     eval_dataset=self.eval_ds,
                     save_path=self.save_dir / "performance_by_age.png",
                 ),